pyproject.toml
[tool.poetry]
name = "filehandling"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.8.4"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
scipy = "^1.13.0"
wave = "^0.0.2"
tensorflow = "^2.16.1"
pillow = "^10.3.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
